backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
sqlalchemy
openai